feat: confirm组件构建

+++ src/views/layout/components/header/header-search/history.vue
@@ -34,6 +34,7 @@ const EMITS_ITEM_CLICK = 'itemClick'
 
 <script setup>
 import { useStore } from 'vuex'
+import { confirm } from '@/libs'
 
 const emits = defineEmits([EMITS_ITEM_CLICK])
 const store = useStore()
@@ -41,10 +42,13 @@ const store = useStore()
  * 删除所有记录
  */
 const onDeleteAllClick = () => {
-  store.commit('search/deleteAllHistory')
-  // confirm('要删除所有历史记录吗？').then(() => {
-  //   store.commit('search/deleteAllHistory')
-  // })
+  confirm('是否删除所有历史记录？')
+    .then(() => {
+      store.commit('search/deleteAllHistory')
+    })
+    .catch(() => {
+      console.log('cancel')
+    })
 }
 
 /**
